* scaffolding(mem): basic scaffolding for mm rewrite


* feat(lib): add rb tree impl


* feat(mem): implement VMA subsys


* test(mem): temp VMA unit test


* refactor(mem): decouple PMM from limine


* scaffold(mem): more mm scaffolds


* feat(mmu): implement mmu driver


* test(mmu): add mmu unit tests


* quality(mmu): improved code quality in mmu


* mv(mmu): move mmu from core/ to mem/


* mv(mem): move tests to mem/test/


* rm(vma): scaffolding warning


* feat(pmm): create PMM


* tests(pmm): create pmm unit tests


* rm(dead): dead code removal


* ref(mem): update memory_managerreferences


* nuke(mem): nuke old memory manager code


* feat(simd): fast memory copies and zeroing using SIMD


* feat(pmm): support slab flags and fast zeroing


* feat(slab): implement slab allocator


* tests(slab): create slab unit tests


* refactor(core): wire up slab allocator and cleanup initrd buffer


* fix(net): add spinlock to network_init


* fix(pty): kill foreground process group on pty destroy


* git: update submodule references


* feat(mmu): add page table cloning and page fault hook


* feat(vmm): implement vmm


* tests(vmm): create vmm unit tests


* feat(lib): add radix tree impl


* feat(pagecache): implement page cache


* tests(pagecache): create page cache unit tests


* feat(fs): add dirty page flusher


* rm(fs): remove shit bootfs and inkernel mkfs


* feat(vfs): implement tmpfs


* feat(disk): add gpt rescanning and refactor fat32


* fix(ac97): fix ac97 driver for new mm


* feat(drivers): update ps2 and i2c drivers


* refactor(net): update socket buffers and e1000 driver


* feat(sys): wire process and syscalls to vmm


* refactor(core): update boot flags and update limine path


* git: update submodule references


---------

+++ drivers/I2C/i2c_lpss.c
@@ -7,13 +7,13 @@
 #include "acpi_i2c.h"
 #include "pci.h"
 #include "platform.h"
-#include "memory_manager.h"
+#include "slab.h"
 #include "paging.h"
 #include "kutils.h"
 #include <string.h>
 
 extern void serial_write(const char *str);
-extern void serial_write_hex(uint32_t val);
+extern void serial_write_hex(uint64_t val);
 extern void serial_write_num(uint64_t num);
 
 #define MAX_LPSS_I2C_CONTROLLERS 8
